Structuring backend (#495)
* split rocsparse_handle.hpp

* split rocsparse_handle.cpp

* structuring trm_info without behaviour change

* fixing test failure related to copy_mat_info

* Fixing default initialization of rocsparse_trm_info.
Removing error prone cast when calling hipMalloc
