+++ library/src/CMakeLists.txt
@@ -25,12 +25,14 @@
 set(rocsparse_source
 
 # Auxiliary
-  src/auxiliary/handle.cpp
   src/auxiliary/rocsparse_auxiliary.cpp
   src/auxiliary/rocsparse_error.cpp
 
 # Common
   src/common/rocsparse_common.cpp
+  src/common/rocsparse_handle.cpp
+  src/common/rocsparse_mat_info.cpp
+  src/common/rocsparse_trm_info.cpp
   src/common/rocsparse_indextype_utils.cpp
   src/common/rocsparse_datatype_utils.cpp
   src/common/rocsparse_one.cpp
@@ -101,6 +103,9 @@ set(rocsparse_source
   src/level2/rocsparse_csrmv_template_lrb.cpp
   src/level2/rocsparse_csrmv_impl.cpp
   src/level2/rocsparse_csrmv.cpp
+  src/level2/rocsparse_csrmv_info.cpp
+  src/level2/rocsparse_csrmv_adaptive_info.cpp
+  src/level2/rocsparse_csrmv_lrb_info.cpp
   src/level2/rocsparse_csrmv_analysis.cpp
   #
   src/level2/rocsparse_cscmv_impl.cpp
@@ -115,6 +120,7 @@ set(rocsparse_source
   src/level2/rocsparse_csrsv_buffer_size.cpp
   src/level2/rocsparse_csrsv_solve.cpp
   src/level2/rocsparse_csritsv.cpp
+  src/level2/rocsparse_csritsv_info.cpp
   src/level2/rocsparse_csritsv_buffer_size.cpp
   src/level2/rocsparse_csritsv_analysis.cpp
   src/level2/rocsparse_csritsv_solve.cpp
@@ -377,6 +383,7 @@ set(rocsparse_source
   src/extra/rocsparse_csrgeam_nnz.cpp
   src/extra/rocsparse_csrgeam_buffer_size.cpp
   #
+  src/extra/rocsparse_csrgemm_info.cpp
   src/extra/rocsparse_csrgemm_scal_buffer_size.cpp
   src/extra/rocsparse_csrgemm_scal_nnz.cpp
   src/extra/rocsparse_csrgemm_scal.cpp
