+++ library/src/level2/rocsparse_csrmv_template_lrb.cpp
@@ -55,12 +55,12 @@ rocsparse_status rocsparse::csrmv_analysis_lrb_template_dispatch(rocsparse_handl
     // Stream
     hipStream_t stream = handle->stream;
 
-    RETURN_IF_HIP_ERROR(rocsparse_hipMallocAsync(
-        (void**)&csrmv_info->lrb.rows_offsets_scratch, sizeof(J) * m, stream));
     RETURN_IF_HIP_ERROR(
-        rocsparse_hipMallocAsync((void**)&csrmv_info->lrb.rows_bins, sizeof(J) * m, stream));
+        rocsparse_hipMallocAsync(&csrmv_info->lrb.rows_offsets_scratch, sizeof(J) * m, stream));
     RETURN_IF_HIP_ERROR(
-        rocsparse_hipMallocAsync((void**)&csrmv_info->lrb.n_rows_bins, sizeof(J) * 32, stream));
+        rocsparse_hipMallocAsync(&csrmv_info->lrb.rows_bins, sizeof(J) * m, stream));
+    RETURN_IF_HIP_ERROR(
+        rocsparse_hipMallocAsync(&csrmv_info->lrb.n_rows_bins, sizeof(J) * 32, stream));
 
     RETURN_IF_HIP_ERROR(
         hipMemsetAsync(csrmv_info->lrb.rows_offsets_scratch, 0, sizeof(J) * m, stream));
@@ -197,7 +197,7 @@ rocsparse_status rocsparse::csrmv_analysis_lrb_template_dispatch(rocsparse_handl
         csrmv_info->lrb.size = max_required_grid;
 
         RETURN_IF_HIP_ERROR(rocsparse_hipMallocAsync(
-            (void**)&csrmv_info->lrb.wg_flags, sizeof(uint32_t) * csrmv_info->lrb.size, stream));
+            &csrmv_info->lrb.wg_flags, sizeof(uint32_t) * csrmv_info->lrb.size, stream));
     }
 
     // Store some pointers to verify correct execution
