+++ library/src/level2/rocsparse_csrsv.cpp
@@ -138,19 +138,19 @@ try
     // Clear csrsv meta data (this includes lower, upper and their transposed equivalents
     if(!rocsparse::check_trm_shared(info, info->csrsv_lower_info))
     {
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::destroy_trm_info(info->csrsv_lower_info));
+        rocsparse::trm_info_t::destroy(info->csrsv_lower_info);
     }
     if(!rocsparse::check_trm_shared(info, info->csrsvt_lower_info))
     {
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::destroy_trm_info(info->csrsvt_lower_info));
+        rocsparse::trm_info_t::destroy(info->csrsvt_lower_info);
     }
     if(!rocsparse::check_trm_shared(info, info->csrsv_upper_info))
     {
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::destroy_trm_info(info->csrsv_upper_info));
+        rocsparse::trm_info_t::destroy(info->csrsv_upper_info);
     }
     if(!rocsparse::check_trm_shared(info, info->csrsvt_upper_info))
     {
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::destroy_trm_info(info->csrsvt_upper_info));
+        rocsparse::trm_info_t::destroy(info->csrsvt_upper_info);
     }
 
     info->csrsv_lower_info  = nullptr;
