+++ library/src/level2/rocsparse_csrsv_analysis.cpp
@@ -44,7 +44,7 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
                                          const T*                  csr_val,
                                          const I*                  csr_row_ptr,
                                          const J*                  csr_col_ind,
-                                         rocsparse_trm_info        info,
+                                         rocsparse::trm_info_t*    info,
                                          J**                       zero_pivot,
                                          void*                     temp_buffer)
 {
@@ -58,8 +58,8 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
     {
         // TODO: this need to be changed.
         // LCOV_EXCL_START
-        if(info->trmt_perm != nullptr || info->trmt_row_ptr != nullptr
-           || info->trmt_col_ind != nullptr)
+        if(info->get_transposed_perm() != nullptr || info->get_transposed_row_ptr() != nullptr
+           || info->get_transposed_col_ind() != nullptr)
         {
             RETURN_IF_ROCSPARSE_ERROR(rocsparse_status_internal_error);
         }
@@ -83,23 +83,32 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
         RETURN_IF_HIP_ERROR(hipMemcpyAsync(
             tmp_work1, csr_col_ind, sizeof(J) * nnz, hipMemcpyDeviceToDevice, stream));
 
+        void* transposed_row_ptr{};
         RETURN_IF_HIP_ERROR(
-            rocsparse_hipMallocAsync((void**)&info->trmt_row_ptr, sizeof(I) * (m + 1), stream));
+            rocsparse_hipMallocAsync(&transposed_row_ptr, sizeof(I) * (m + 1), stream));
+
+        info->set_transposed_row_ptr(transposed_row_ptr);
 
         if(nnz > 0)
         {
+            void* transposed_perm{};
+            void* transposed_col_ind{};
+
             RETURN_IF_HIP_ERROR(
-                rocsparse_hipMallocAsync((void**)&info->trmt_perm, sizeof(I) * nnz, stream));
+                rocsparse_hipMallocAsync(&transposed_perm, sizeof(I) * nnz, stream));
+            info->set_transposed_perm(transposed_perm);
+
             RETURN_IF_HIP_ERROR(
-                rocsparse_hipMallocAsync((void**)&info->trmt_col_ind, sizeof(J) * nnz, stream));
+                rocsparse_hipMallocAsync(&transposed_col_ind, sizeof(J) * nnz, stream));
+            info->set_transposed_col_ind(transposed_col_ind);
 
             // Create identity permutation
             RETURN_IF_ROCSPARSE_ERROR(
-                rocsparse::create_identity_permutation_template(handle, nnz, (I*)info->trmt_perm));
+                rocsparse::create_identity_permutation_template(handle, nnz, (I*)transposed_perm));
 
             // Stable sort COO by columns
-            rocsparse::primitives::double_buffer<J> keys(tmp_work1, (J*)info->trmt_col_ind);
-            rocsparse::primitives::double_buffer<I> vals((I*)info->trmt_perm, tmp_work2);
+            rocsparse::primitives::double_buffer<J> keys(tmp_work1, (J*)transposed_col_ind);
+            rocsparse::primitives::double_buffer<I> vals((I*)transposed_perm, tmp_work2);
 
             uint32_t startbit = 0;
             uint32_t endbit   = rocsparse::clz(m);
@@ -111,9 +120,9 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
                 handle, keys, vals, nnz, startbit, endbit, rocprim_size, rocprim_buffer));
 
             // Copy permutation vector, if not already available
-            if(vals.current() != info->trmt_perm)
+            if(vals.current() != transposed_perm)
             {
-                RETURN_IF_HIP_ERROR(hipMemcpyAsync(info->trmt_perm,
+                RETURN_IF_HIP_ERROR(hipMemcpyAsync(transposed_perm,
                                                    vals.current(),
                                                    sizeof(I) * nnz,
                                                    hipMemcpyDeviceToDevice,
@@ -122,7 +131,7 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
 
             // Create column pointers
             RETURN_IF_ROCSPARSE_ERROR(rocsparse::coo2csr_template(
-                handle, keys.current(), nnz, m, (I*)info->trmt_row_ptr, descr->base));
+                handle, keys.current(), nnz, m, (I*)transposed_row_ptr, descr->base));
 
             // Create row indices
             RETURN_IF_ROCSPARSE_ERROR(
@@ -132,14 +141,14 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
             RETURN_IF_ROCSPARSE_ERROR(rocsparse::gthr_template(handle,
                                                                nnz,
                                                                tmp_work1,
-                                                               (J*)info->trmt_col_ind,
-                                                               (const I*)info->trmt_perm,
+                                                               (J*)transposed_col_ind,
+                                                               (const I*)transposed_perm,
                                                                rocsparse_index_base_zero));
         }
         else
         {
             RETURN_IF_ROCSPARSE_ERROR(rocsparse::valset(
-                handle, m + 1, static_cast<I>(descr->base), (I*)info->trmt_row_ptr));
+                handle, m + 1, static_cast<I>(descr->base), (I*)transposed_row_ptr));
         }
     }
 
@@ -170,18 +179,20 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
     void* rocprim_buffer = reinterpret_cast<void*>(ptr);
 
     // Allocate buffer to hold diagonal entry point
-    RETURN_IF_HIP_ERROR(
-        rocsparse_hipMallocAsync((void**)&info->trm_diag_ind, sizeof(I) * m, stream));
+    I* diag_ind{};
+    RETURN_IF_HIP_ERROR(rocsparse_hipMallocAsync(&diag_ind, sizeof(I) * m, stream));
+    info->set_diag_ind(diag_ind);
 
     // Allocate buffer to hold zero pivot
     if(*zero_pivot == nullptr)
     {
-        RETURN_IF_HIP_ERROR(rocsparse_hipMallocAsync((void**)zero_pivot, sizeof(J), stream));
+        RETURN_IF_HIP_ERROR(rocsparse_hipMallocAsync(zero_pivot, sizeof(J), stream));
     }
 
+    J* row_map{};
     // Allocate buffer to hold row map
-    RETURN_IF_HIP_ERROR(rocsparse_hipMallocAsync((void**)&info->row_map, sizeof(J) * m, stream));
-
+    RETURN_IF_HIP_ERROR(rocsparse_hipMallocAsync(&row_map, sizeof(J) * m, stream));
+    info->set_row_map(row_map);
     // Initialize zero pivot
     RETURN_IF_ROCSPARSE_ERROR(
         rocsparse::assign_async(*zero_pivot, std::numeric_limits<J>::max(), stream));
@@ -211,7 +222,7 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
                     m,
                     csr_row_ptr,
                     csr_col_ind,
-                    (I*)info->trm_diag_ind,
+                    diag_ind,
                     done_array,
                     d_max_nnz,
                     (J*)*zero_pivot,
@@ -229,7 +240,7 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
                     m,
                     csr_row_ptr,
                     csr_col_ind,
-                    (I*)info->trm_diag_ind,
+                    diag_ind,
                     done_array,
                     d_max_nnz,
                     (J*)*zero_pivot,
@@ -254,7 +265,7 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
                         m,
                         csr_row_ptr,
                         csr_col_ind,
-                        (I*)info->trm_diag_ind,
+                        diag_ind,
                         done_array,
                         d_max_nnz,
                         (J*)*zero_pivot,
@@ -272,7 +283,7 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
                         m,
                         csr_row_ptr,
                         csr_col_ind,
-                        (I*)info->trm_diag_ind,
+                        diag_ind,
                         done_array,
                         d_max_nnz,
                         (J*)*zero_pivot,
@@ -296,7 +307,7 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
                         m,
                         csr_row_ptr,
                         csr_col_ind,
-                        (I*)info->trm_diag_ind,
+                        diag_ind,
                         done_array,
                         d_max_nnz,
                         (J*)*zero_pivot,
@@ -314,7 +325,7 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
                         m,
                         csr_row_ptr,
                         csr_col_ind,
-                        (I*)info->trm_diag_ind,
+                        diag_ind,
                         done_array,
                         d_max_nnz,
                         (J*)*zero_pivot,
@@ -339,9 +350,9 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
                     0,
                     stream,
                     m,
-                    (const I*)info->trmt_row_ptr,
-                    (const J*)info->trmt_col_ind,
-                    (I*)info->trm_diag_ind,
+                    (const I*)info->get_transposed_row_ptr(),
+                    (const J*)info->get_transposed_col_ind(),
+                    diag_ind,
                     done_array,
                     d_max_nnz,
                     (J*)*zero_pivot,
@@ -357,9 +368,9 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
                     0,
                     stream,
                     m,
-                    (const I*)info->trmt_row_ptr,
-                    (const J*)info->trmt_col_ind,
-                    (I*)info->trm_diag_ind,
+                    (const I*)info->get_transposed_row_ptr(),
+                    (const J*)info->get_transposed_col_ind(),
+                    diag_ind,
                     done_array,
                     d_max_nnz,
                     (J*)*zero_pivot,
@@ -382,9 +393,9 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
                         0,
                         stream,
                         m,
-                        (const I*)info->trmt_row_ptr,
-                        (const J*)info->trmt_col_ind,
-                        (I*)info->trm_diag_ind,
+                        (const I*)info->get_transposed_row_ptr(),
+                        (const J*)info->get_transposed_col_ind(),
+                        diag_ind,
                         done_array,
                         d_max_nnz,
                         (J*)*zero_pivot,
@@ -400,9 +411,9 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
                         0,
                         stream,
                         m,
-                        (const I*)info->trmt_row_ptr,
-                        (const J*)info->trmt_col_ind,
-                        (I*)info->trm_diag_ind,
+                        (const I*)info->get_transposed_row_ptr(),
+                        (const J*)info->get_transposed_col_ind(),
+                        diag_ind,
                         done_array,
                         d_max_nnz,
                         (J*)*zero_pivot,
@@ -424,9 +435,9 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
                         0,
                         stream,
                         m,
-                        (const I*)info->trmt_row_ptr,
-                        (const J*)info->trmt_col_ind,
-                        (I*)info->trm_diag_ind,
+                        (const I*)info->get_transposed_row_ptr(),
+                        (const J*)info->get_transposed_col_ind(),
+                        diag_ind,
                         done_array,
                         d_max_nnz,
                         (J*)*zero_pivot,
@@ -442,9 +453,9 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
                         0,
                         stream,
                         m,
-                        (const I*)info->trmt_row_ptr,
-                        (const J*)info->trmt_col_ind,
-                        (I*)info->trm_diag_ind,
+                        (const I*)info->get_transposed_row_ptr(),
+                        (const J*)info->get_transposed_col_ind(),
+                        diag_ind,
                         done_array,
                         d_max_nnz,
                         (J*)*zero_pivot,
@@ -463,9 +474,10 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
 #undef CSRSV_DIM
 
     // Post processing
+    int64_t max_nnz;
     RETURN_IF_HIP_ERROR(
-        hipMemcpyAsync(&info->max_nnz, d_max_nnz, sizeof(I), hipMemcpyDeviceToHost, stream));
-
+        hipMemcpyAsync(&max_nnz, d_max_nnz, sizeof(I), hipMemcpyDeviceToHost, stream));
+    info->set_max_nnz(max_nnz);
     // Wait for host transfer to finish
     RETURN_IF_HIP_ERROR(hipStreamSynchronize(stream));
 
@@ -478,34 +490,39 @@ rocsparse_status rocsparse::trm_analysis(rocsparse_handle          handle,
     uint32_t endbit   = rocsparse::clz(m);
 
     rocsparse::primitives::double_buffer<int> keys(done_array, workspace2);
-    rocsparse::primitives::double_buffer<J>   vals(workspace, (J*)info->row_map);
+    rocsparse::primitives::double_buffer<J>   vals(workspace, row_map);
 
     RETURN_IF_ROCSPARSE_ERROR((rocsparse::primitives::radix_sort_pairs_buffer_size<int, J>(
         handle, m, startbit, endbit, &rocprim_size)));
     RETURN_IF_ROCSPARSE_ERROR(rocsparse::primitives::radix_sort_pairs(
         handle, keys, vals, m, startbit, endbit, rocprim_size, rocprim_buffer));
 
-    if(vals.current() != info->row_map)
+    if(vals.current() != row_map)
     {
         RETURN_IF_HIP_ERROR(hipMemcpyAsync(
-            info->row_map, vals.current(), sizeof(J) * m, hipMemcpyDeviceToDevice, stream));
+            row_map, vals.current(), sizeof(J) * m, hipMemcpyDeviceToDevice, stream));
     }
 
     // Store some pointers to verify correct execution
-    info->m           = m;
-    info->nnz         = nnz;
-    info->descr       = descr;
-    info->trm_row_ptr = (trans == rocsparse_operation_none) ? csr_row_ptr : info->trmt_row_ptr;
-    info->trm_col_ind = (trans == rocsparse_operation_none) ? csr_col_ind : info->trmt_col_ind;
 
-    info->index_type_I = (sizeof(I) == sizeof(uint16_t))
-                             ? rocsparse_indextype_u16
-                             : ((sizeof(I) == sizeof(int32_t)) ? rocsparse_indextype_i32
-                                                               : rocsparse_indextype_i64);
-    info->index_type_J = (sizeof(J) == sizeof(uint16_t))
-                             ? rocsparse_indextype_u16
-                             : ((sizeof(J) == sizeof(int32_t)) ? rocsparse_indextype_i32
-                                                               : rocsparse_indextype_i64);
+    info->set_m(m);
+    info->set_nnz(nnz);
+    info->set_descr(descr);
+
+    info->set_row_ptr((trans == rocsparse_operation_none) ? csr_row_ptr
+                                                          : info->get_transposed_row_ptr());
+    info->set_col_ind((trans == rocsparse_operation_none) ? csr_col_ind
+                                                          : info->get_transposed_col_ind());
+
+    info->set_offset_indextype(
+        (sizeof(I) == sizeof(uint16_t))
+            ? rocsparse_indextype_u16
+            : ((sizeof(I) == sizeof(int32_t)) ? rocsparse_indextype_i32 : rocsparse_indextype_i64));
+
+    info->set_index_indextype(
+        (sizeof(J) == sizeof(uint16_t))
+            ? rocsparse_indextype_u16
+            : ((sizeof(J) == sizeof(int32_t)) ? rocsparse_indextype_i32 : rocsparse_indextype_i64));
 
     return rocsparse_status_success;
 }
@@ -633,29 +650,24 @@ rocsparse_status rocsparse::csrsv_analysis_template(rocsparse_handle          ha
         // User is explicitly asking to force a re-analysis, or no valid data has been
         // found to be re-used.
 
-        // Clear csrsv info
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::destroy_trm_info((trans == rocsparse_operation_none)
-                                                                  ? info->csrsv_upper_info
-                                                                  : info->csrsvt_upper_info));
+        rocsparse::trm_info_t** p_trm_info = (trans == rocsparse_operation_none)
+                                                 ? &info->csrsv_upper_info
+                                                 : &info->csrsvt_upper_info;
 
-        // Create csrsv info
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::create_trm_info((trans == rocsparse_operation_none)
-                                                                 ? &info->csrsv_upper_info
-                                                                 : &info->csrsvt_upper_info));
+        rocsparse::trm_info_t::recreate(p_trm_info);
 
         // Perform analysis
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::trm_analysis(
-            handle,
-            trans,
-            m,
-            nnz,
-            descr,
-            csr_val,
-            csr_row_ptr,
-            csr_col_ind,
-            (trans == rocsparse_operation_none) ? info->csrsv_upper_info : info->csrsvt_upper_info,
-            (J**)&info->zero_pivot,
-            temp_buffer));
+        RETURN_IF_ROCSPARSE_ERROR(rocsparse::trm_analysis(handle,
+                                                          trans,
+                                                          m,
+                                                          nnz,
+                                                          descr,
+                                                          csr_val,
+                                                          csr_row_ptr,
+                                                          csr_col_ind,
+                                                          p_trm_info[0],
+                                                          (J**)&info->zero_pivot,
+                                                          temp_buffer));
     }
     else
     {
@@ -719,29 +731,24 @@ rocsparse_status rocsparse::csrsv_analysis_template(rocsparse_handle          ha
         // User is explicitly asking to force a re-analysis, or no valid data has been
         // found to be re-used.
 
-        // Clear csrsv info
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::destroy_trm_info((trans == rocsparse_operation_none)
-                                                                  ? info->csrsv_lower_info
-                                                                  : info->csrsvt_lower_info));
+        rocsparse::trm_info_t** p_trm_info = (trans == rocsparse_operation_none)
+                                                 ? &info->csrsv_lower_info
+                                                 : &info->csrsvt_lower_info;
 
-        // Create csrsv info
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::create_trm_info((trans == rocsparse_operation_none)
-                                                                 ? &info->csrsv_lower_info
-                                                                 : &info->csrsvt_lower_info));
+        rocsparse::trm_info_t::recreate(p_trm_info);
 
         // Perform analysis
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::trm_analysis(
-            handle,
-            trans,
-            m,
-            nnz,
-            descr,
-            csr_val,
-            csr_row_ptr,
-            csr_col_ind,
-            (trans == rocsparse_operation_none) ? info->csrsv_lower_info : info->csrsvt_lower_info,
-            (J**)&info->zero_pivot,
-            temp_buffer));
+        RETURN_IF_ROCSPARSE_ERROR(rocsparse::trm_analysis(handle,
+                                                          trans,
+                                                          m,
+                                                          nnz,
+                                                          descr,
+                                                          csr_val,
+                                                          csr_row_ptr,
+                                                          csr_col_ind,
+                                                          p_trm_info[0],
+                                                          (J**)&info->zero_pivot,
+                                                          temp_buffer));
     }
 
     return rocsparse_status_success;
@@ -756,7 +763,7 @@ rocsparse_status rocsparse::csrsv_analysis_template(rocsparse_handle          ha
                                                       const TTYPE*              csr_val,     \
                                                       const ITYPE*              csr_row_ptr, \
                                                       const JTYPE*              csr_col_ind, \
-                                                      rocsparse_trm_info        info,        \
+                                                      rocsparse::trm_info_t*    info,        \
                                                       JTYPE**                   zero_pivot,  \
                                                       void*                     temp_buffer);
