+++ library/src/level3/csrsm_device.h
@@ -39,7 +39,7 @@ namespace rocsparse
                                            T* __restrict__ B,
                                            int64_t ldb,
                                            int* __restrict__ done_array,
-                                           J* __restrict__ map,
+                                           const J* __restrict__ map,
                                            J* __restrict__ zero_pivot,
                                            rocsparse_index_base idx_base,
                                            rocsparse_fill_mode  fill_mode,
