+++ library/src/level3/rocsparse_csrsm_analysis.cpp
@@ -123,15 +123,11 @@ rocsparse_status rocsparse::csrsm_analysis_core(rocsparse_handle          handle
         // User is explicitly asking to force a re-analysis, or no valid data has been
         // found to be re-used
 
-        // Clear csrsm info
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::destroy_trm_info((trans_A == rocsparse_operation_none)
-                                                                  ? info->csrsm_upper_info
-                                                                  : info->csrsmt_upper_info));
+        rocsparse::trm_info_t** p_trm_info = (trans_A == rocsparse_operation_none)
+                                                 ? &info->csrsm_upper_info
+                                                 : &info->csrsmt_upper_info;
 
-        // Create csrsm info
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::create_trm_info((trans_A == rocsparse_operation_none)
-                                                                 ? &info->csrsm_upper_info
-                                                                 : &info->csrsmt_upper_info));
+        rocsparse::trm_info_t::recreate(p_trm_info);
 
         // Perform analysis
         RETURN_IF_ROCSPARSE_ERROR(rocsparse::trm_analysis(handle,
@@ -142,9 +138,7 @@ rocsparse_status rocsparse::csrsm_analysis_core(rocsparse_handle          handle
                                                           csr_val,
                                                           csr_row_ptr,
                                                           csr_col_ind,
-                                                          (trans_A == rocsparse_operation_none)
-                                                              ? info->csrsm_upper_info
-                                                              : info->csrsmt_upper_info,
+                                                          p_trm_info[0],
                                                           (J**)&info->zero_pivot,
                                                           temp_buffer));
     }
@@ -212,15 +206,11 @@ rocsparse_status rocsparse::csrsm_analysis_core(rocsparse_handle          handle
         // User is explicitly asking to force a re-analysis, or no valid data has been
         // found to be re-used
 
-        // Clear csrsm info
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::destroy_trm_info((trans_A == rocsparse_operation_none)
-                                                                  ? info->csrsm_lower_info
-                                                                  : info->csrsmt_lower_info));
+        rocsparse::trm_info_t** p_trm_info = (trans_A == rocsparse_operation_none)
+                                                 ? &info->csrsm_lower_info
+                                                 : &info->csrsmt_lower_info;
 
-        // Create csrsm info
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::create_trm_info((trans_A == rocsparse_operation_none)
-                                                                 ? &info->csrsm_lower_info
-                                                                 : &info->csrsmt_lower_info));
+        rocsparse::trm_info_t::recreate(p_trm_info);
 
         // Perform analysis
         RETURN_IF_ROCSPARSE_ERROR(rocsparse::trm_analysis(handle,
@@ -231,9 +221,7 @@ rocsparse_status rocsparse::csrsm_analysis_core(rocsparse_handle          handle
                                                           csr_val,
                                                           csr_row_ptr,
                                                           csr_col_ind,
-                                                          (trans_A == rocsparse_operation_none)
-                                                              ? info->csrsm_lower_info
-                                                              : info->csrsmt_lower_info,
+                                                          p_trm_info[0],
                                                           (J**)&info->zero_pivot,
                                                           temp_buffer));
     }
