+++ library/src/precond/rocsparse_csrilu0.cpp
@@ -179,7 +179,7 @@ namespace rocsparse
         RETURN_IF_HIP_ERROR(hipMemsetAsync(d_done_array, 0, sizeof(int) * m, stream));
 
         // Max nnz per row
-        rocsparse_int max_nnz = info->csrilu0_info->max_nnz;
+        const rocsparse_int max_nnz = info->csrilu0_info->get_max_nnz();
 
         // Determine gcnArch and ASIC revision
         const std::string gcn_arch_name = rocsparse::handle_get_arch_name(handle);
@@ -203,9 +203,9 @@ namespace rocsparse
                 csr_row_ptr,
                 csr_col_ind,
                 csr_val,
-                (rocsparse_int*)info->csrilu0_info->trm_diag_ind,
+                (const rocsparse_int*)info->csrilu0_info->get_diag_ind(),
                 d_done_array,
-                (rocsparse_int*)info->csrilu0_info->row_map,
+                (const rocsparse_int*)info->csrilu0_info->get_row_map(),
                 (rocsparse_int*)info->zero_pivot,
                 (rocsparse_int*)info->singular_pivot,
                 info->singular_tol,
@@ -233,9 +233,9 @@ namespace rocsparse
                         csr_row_ptr,
                         csr_col_ind,
                         csr_val,
-                        (rocsparse_int*)info->csrilu0_info->trm_diag_ind,
+                        (const rocsparse_int*)info->csrilu0_info->get_diag_ind(),
                         d_done_array,
-                        (rocsparse_int*)info->csrilu0_info->row_map,
+                        (const rocsparse_int*)info->csrilu0_info->get_row_map(),
                         (rocsparse_int*)info->zero_pivot,
                         (rocsparse_int*)info->singular_pivot,
                         info->singular_tol,
@@ -259,9 +259,9 @@ namespace rocsparse
                         csr_row_ptr,
                         csr_col_ind,
                         csr_val,
-                        (rocsparse_int*)info->csrilu0_info->trm_diag_ind,
+                        (const rocsparse_int*)info->csrilu0_info->get_diag_ind(),
                         d_done_array,
-                        (rocsparse_int*)info->csrilu0_info->row_map,
+                        (const rocsparse_int*)info->csrilu0_info->get_row_map(),
                         (rocsparse_int*)info->zero_pivot,
                         (rocsparse_int*)info->singular_pivot,
                         info->singular_tol,
@@ -285,9 +285,9 @@ namespace rocsparse
                         csr_row_ptr,
                         csr_col_ind,
                         csr_val,
-                        (rocsparse_int*)info->csrilu0_info->trm_diag_ind,
+                        (const rocsparse_int*)info->csrilu0_info->get_diag_ind(),
                         d_done_array,
-                        (rocsparse_int*)info->csrilu0_info->row_map,
+                        (const rocsparse_int*)info->csrilu0_info->get_row_map(),
                         (rocsparse_int*)info->zero_pivot,
                         (rocsparse_int*)info->singular_pivot,
                         info->singular_tol,
@@ -311,9 +311,9 @@ namespace rocsparse
                         csr_row_ptr,
                         csr_col_ind,
                         csr_val,
-                        (rocsparse_int*)info->csrilu0_info->trm_diag_ind,
+                        (const rocsparse_int*)info->csrilu0_info->get_diag_ind(),
                         d_done_array,
-                        (rocsparse_int*)info->csrilu0_info->row_map,
+                        (const rocsparse_int*)info->csrilu0_info->get_row_map(),
                         (rocsparse_int*)info->zero_pivot,
                         (rocsparse_int*)info->singular_pivot,
                         info->singular_tol,
@@ -337,9 +337,9 @@ namespace rocsparse
                         csr_row_ptr,
                         csr_col_ind,
                         csr_val,
-                        (rocsparse_int*)info->csrilu0_info->trm_diag_ind,
+                        (const rocsparse_int*)info->csrilu0_info->get_diag_ind(),
                         d_done_array,
-                        (rocsparse_int*)info->csrilu0_info->row_map,
+                        (const rocsparse_int*)info->csrilu0_info->get_row_map(),
                         (rocsparse_int*)info->zero_pivot,
                         (rocsparse_int*)info->singular_pivot,
                         info->singular_tol,
@@ -363,9 +363,9 @@ namespace rocsparse
                         csr_row_ptr,
                         csr_col_ind,
                         csr_val,
-                        (rocsparse_int*)info->csrilu0_info->trm_diag_ind,
+                        (const rocsparse_int*)info->csrilu0_info->get_diag_ind(),
                         d_done_array,
-                        (rocsparse_int*)info->csrilu0_info->row_map,
+                        (const rocsparse_int*)info->csrilu0_info->get_row_map(),
                         (rocsparse_int*)info->zero_pivot,
                         (rocsparse_int*)info->singular_pivot,
                         info->singular_tol,
@@ -392,9 +392,9 @@ namespace rocsparse
                         csr_row_ptr,
                         csr_col_ind,
                         csr_val,
-                        (rocsparse_int*)info->csrilu0_info->trm_diag_ind,
+                        (const rocsparse_int*)info->csrilu0_info->get_diag_ind(),
                         d_done_array,
-                        (rocsparse_int*)info->csrilu0_info->row_map,
+                        (const rocsparse_int*)info->csrilu0_info->get_row_map(),
                         (rocsparse_int*)info->zero_pivot,
                         (rocsparse_int*)info->singular_pivot,
                         info->singular_tol,
@@ -418,9 +418,9 @@ namespace rocsparse
                         csr_row_ptr,
                         csr_col_ind,
                         csr_val,
-                        (rocsparse_int*)info->csrilu0_info->trm_diag_ind,
+                        (const rocsparse_int*)info->csrilu0_info->get_diag_ind(),
                         d_done_array,
-                        (rocsparse_int*)info->csrilu0_info->row_map,
+                        (const rocsparse_int*)info->csrilu0_info->get_row_map(),
                         (rocsparse_int*)info->zero_pivot,
                         (rocsparse_int*)info->singular_pivot,
                         info->singular_tol,
@@ -444,9 +444,9 @@ namespace rocsparse
                         csr_row_ptr,
                         csr_col_ind,
                         csr_val,
-                        (rocsparse_int*)info->csrilu0_info->trm_diag_ind,
+                        (const rocsparse_int*)info->csrilu0_info->get_diag_ind(),
                         d_done_array,
-                        (rocsparse_int*)info->csrilu0_info->row_map,
+                        (const rocsparse_int*)info->csrilu0_info->get_row_map(),
                         (rocsparse_int*)info->zero_pivot,
                         (rocsparse_int*)info->singular_pivot,
                         info->singular_tol,
@@ -470,9 +470,9 @@ namespace rocsparse
                         csr_row_ptr,
                         csr_col_ind,
                         csr_val,
-                        (rocsparse_int*)info->csrilu0_info->trm_diag_ind,
+                        (const rocsparse_int*)info->csrilu0_info->get_diag_ind(),
                         d_done_array,
-                        (rocsparse_int*)info->csrilu0_info->row_map,
+                        (const rocsparse_int*)info->csrilu0_info->get_row_map(),
                         (rocsparse_int*)info->zero_pivot,
                         (rocsparse_int*)info->singular_pivot,
                         info->singular_tol,
@@ -496,9 +496,9 @@ namespace rocsparse
                         csr_row_ptr,
                         csr_col_ind,
                         csr_val,
-                        (rocsparse_int*)info->csrilu0_info->trm_diag_ind,
+                        (const rocsparse_int*)info->csrilu0_info->get_diag_ind(),
                         d_done_array,
-                        (rocsparse_int*)info->csrilu0_info->row_map,
+                        (const rocsparse_int*)info->csrilu0_info->get_row_map(),
                         (rocsparse_int*)info->zero_pivot,
                         (rocsparse_int*)info->singular_pivot,
                         info->singular_tol,
@@ -522,9 +522,9 @@ namespace rocsparse
                         csr_row_ptr,
                         csr_col_ind,
                         csr_val,
-                        (rocsparse_int*)info->csrilu0_info->trm_diag_ind,
+                        (const rocsparse_int*)info->csrilu0_info->get_diag_ind(),
                         d_done_array,
-                        (rocsparse_int*)info->csrilu0_info->row_map,
+                        (const rocsparse_int*)info->csrilu0_info->get_row_map(),
                         (rocsparse_int*)info->zero_pivot,
                         (rocsparse_int*)info->singular_pivot,
                         info->singular_tol,
@@ -829,7 +829,7 @@ try
     // If meta data is not shared, delete it
     if(!rocsparse::check_trm_shared(info, info->csrilu0_info))
     {
-        RETURN_IF_ROCSPARSE_ERROR(rocsparse::destroy_trm_info(info->csrilu0_info));
+        rocsparse::trm_info_t::destroy(info->csrilu0_info);
     }
 
     info->csrilu0_info = nullptr;
